mtd: nand: Add support for NAND partitions
Add support for NAND partitions, and indirectly for per partition ECC
config, and also per partiton random seed support for the upcoming
randomizer support.

This is necessary to be able to use different ECC / randomizer settings for
the parts of the NAND which are read directly by a bootrom (which has a
fixed ECC / random seed setting) and the generic data part of the NAND for
which we often want a stronger ECC and / or random seed.

Provide helper functions to add/delete/allocate nand partitions.
NAND core code now make use of the partition specific nand_ecc_ctrl struct
(if available) when doing read/write operations.

Signed-off-by: Boris BREZILLON <boris.brezillon@free-electrons.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>