ARM: dts: sun5i: Enable USB DRC on A10s OLinuxIno Micro
Enable the otg/drc usb controller on the A10s OLinuxIno Micro.

Note the A10s OlinuxIno Micro always has some voltage on its otg power pin,
even if the usb-vbus0-regulator is disabled, the leaked voltage is enough to
make vbus-det always report 1, so we do not use vbus-det on this board.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>